backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "dj-rest-auth>=7.0.1",
    "django>=5.2.6",
    "django-allauth>=65.11.2",
    "django-axes>=8.0.0",
    "django-environ>=0.12.0",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "faker>=37.8.0",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-django>=4.11.1",
    "ruff>=0.12.12",
]
